basic but bad peephole for reloads + move out simple register allocator
