+++ src/codegen-x86-ir.c
@@ -11,6 +11,8 @@
 #include "ir.h"
 #include "ir-fold.h"
 #include "ir-mem2reg.h"
+#include "ir-peephole.h"
+#include "ir-regalloc.h"
 #include "ir-types.h"
 #include "list.h"
 #include "prsutil.h"
@@ -29,7 +31,7 @@ extern IrValue *irFnGetVar(IrFunction *func, u32 lvar_id);
  * its tmp's slot.
  *
  * On top of that we run a tiny "fusion" pre-pass that sets two flag bits on
- * each IrInstr (IRCG_FUSE_TO_NEXT, IRCG_R1_IN_RAX) when a single-use def is
+ * each IrInstr (IRCG_FUSE_TO_NEXT, IRCG_R1_IN_REG) when a single-use def is
  * immediately consumed as the next instruction's first %rax source. The
  * emitter then skips the spill on the def and the reload on the use, and
  * the slot allocator drops the (unused) stack slot. This collapses the bulk
@@ -41,246 +43,29 @@ extern IrValue *irFnGetVar(IrFunction *func, u32 lvar_id);
  * one place.
  */
 
-/* Codegen-private bits stored on IrInstr.flags. ir.c initialises flags to 0
- * and no other code (ir-fold, ir-debug, ARM64) reads or writes it, so we own
- * these bits exclusively during codegen. */
-#define IRCG_FUSE_TO_NEXT (1u << 0)
-#define IRCG_R1_IN_RAX    (1u << 1)
-/* Set on an IR_PHI when the value can live in %rax across the block boundary
- * (the only phi at this block's head and all predecessors arrive via IR_JMP,
- * not IR_BR). When set, predecessors materialise the incoming value into
- * %rax before their jmp; the phi itself emits nothing; the slot allocator
- * skips the phi's dst. */
-#define IRCG_PHI_IN_RAX   (1u << 2)
+/* IRCG_* flag macros and the IrRaCtx live in ir-regalloc.h - they're
+ * shared with the peephole pass which sets the same bits the codegen
+ * here reads. */
 
 typedef struct IrCgCtx {
     Cctrl *cc;
     AoStr *buf;
-    IrFunction *func;
-    /* Map<u32 IrValue.var.id -> int loff>. loff is sign-extended on read. */
-    Map *id_to_loff;
-    /* Bytes added to %rsp beyond what asmFunctionInit reserved, for IR tmps. */
-    int extra_stack;
-    /* Unique block-label prefix per function so multiple slice functions
-     * compiled in the same translation unit don't collide. */
+    /* Slot offset map + extra-stack counter + IR function. The
+     * regalloc helpers in ir-regalloc.c take an IrRaCtx*; the codegen
+     * passes `&ctx->ra` when calling into them. */
+    IrRaCtx ra;
+    /* Unique block-label prefix per function so multiple slice
+     * functions compiled in the same translation unit don't collide. */
     int func_uid;
-    /* Set during the per-block emission loop so terminator codegen knows
-     * who's emitting and which block (if any) follows in layout order. */
+    /* Set during the per-block emission loop so terminator codegen
+     * knows who's emitting and which block (if any) follows in layout
+     * order. */
     IrBlock *cur_block;
     IrBlock *next_block;
 } IrCgCtx;
 
 static int ircg_func_seq = 0;
 
-static void irCgSetLoff(IrCgCtx *ctx, u32 var_id, int loff) {
-    /* mapAddIntOrErr fails on duplicate; use it as an assert for distinctness. */
-    int ok = mapAddIntOrErr(ctx->id_to_loff, var_id, (void *)(intptr_t)loff);
-    if (!ok) {
-        loggerPanic("ir-cg: duplicate loff mapping for var.id=%u\n", var_id);
-    }
-}
-
-static int irCgGetLoff(IrCgCtx *ctx, IrValue *val) {
-    assert(val != NULL);
-    if (!mapHasInt(ctx->id_to_loff, val->as.var.id)) {
-        loggerPanic("ir-cg: no slot for var.id=%u kind=%d\n",
-                    val->as.var.id, val->kind);
-    }
-    return (int)(intptr_t)mapGetInt(ctx->id_to_loff, val->as.var.id);
-}
-
-/* Reserve a fresh stack slot for an SSA temp. */
-static void irCgAllocTmp(IrCgCtx *ctx, IrValue *val, int starting_offset) {
-    if (!val || val->kind != IR_VAL_TMP) return;
-    if (mapHasInt(ctx->id_to_loff, val->as.var.id)) return;
-    ctx->extra_stack += 8;
-    irCgSetLoff(ctx, val->as.var.id, -(starting_offset + ctx->extra_stack));
-}
-
-/* Bind every variable known to the IR (params + alloca'd locals) to its
- * AST-driven loff. Walk the AST function's params and locals lists; the
- * eligibility predicate guarantees they're plain AST_LVAR ints. */
-static void irCgBindAstLoffs(IrCgCtx *ctx, Ast *ast_func) {
-    /* Hidden out-pointer param for struct-by-value return: the layout
-     * pass stashed its loff on `ast_func->loff` and the IR-side
-     * IR_VAL_PARAM lives on `ctx->func->return_value`. */
-    AstType *rt = ast_func->type ? ast_func->type->rettype : NULL;
-    int has_hidden_out_ptr = rt &&
-        (rt->kind == AST_TYPE_CLASS || rt->kind == AST_TYPE_UNION) &&
-        !rt->is_intrinsic && rt->size > 0;
-    if (has_hidden_out_ptr && ctx->func->return_value) {
-        irCgSetLoff(ctx, ctx->func->return_value->as.var.id,
-                    ast_func->loff);
-    }
-    if (ast_func->params) {
-        for (u64 i = 0; i < ast_func->params->size; ++i) {
-            Ast *p = vecGet(Ast *, ast_func->params, i);
-            if (p->kind == AST_VAR_ARGS) {
-                /* Two slots, two lvar_ids — argc and argv. The layout
-                 * pass set their loffs above. */
-                IrValue *cv = irFnGetVar(ctx->func, p->argc->lvar_id);
-                if (cv) irCgSetLoff(ctx, cv->as.var.id, p->argc->loff);
-                IrValue *vv = irFnGetVar(ctx->func, p->argv->lvar_id);
-                if (vv) irCgSetLoff(ctx, vv->as.var.id, p->argv->loff);
-                continue;
-            }
-            u32 pid;
-            if (p->kind == AST_DEFAULT_PARAM) pid = p->declvar->lvar_id;
-            else if (p->kind == AST_FUNPTR)   pid = p->fn_ptr_id;
-            else                              pid = p->lvar_id;
-            IrValue *iv = irFnGetVar(ctx->func, pid);
-            if (iv) irCgSetLoff(ctx, iv->as.var.id, p->loff);
-        }
-    }
-    listForEach(ast_func->locals) {
-        Ast *l = (Ast *)it->value;
-        u32 lid;
-        if (l->kind == AST_DEFAULT_PARAM) lid = l->declvar->lvar_id;
-        else if (l->kind == AST_FUNPTR)   lid = l->fn_ptr_id;
-        else                              lid = l->lvar_id;
-        IrValue *iv = irFnGetVar(ctx->func, lid);
-        if (iv) irCgSetLoff(ctx, iv->as.var.id, l->loff);
-    }
-}
-
-/* Allocate slots only for tmps the emitter is actually going to touch.
- * Operand roles per opcode:
- *   - "spill dst"    (write rax to slot) — skip if IRCG_FUSE_TO_NEXT.
- *   - "load r1"      (read slot to rax)  — skip if IRCG_R1_IN_RAX.
- *   - "address dst"  (offset literal)    — always needs a slot.
- *   - "load r2"      (read slot to rcx)  — always needs a slot.
- * Fused (dst, r1) pairs are never spilled or read, so their tmp ids
- * legitimately have no slot. */
-static void irCgAllocOperandsForInstr(IrCgCtx *ctx, IrInstr *I, int start) {
-    int spill_dst   = !(I->flags & IRCG_FUSE_TO_NEXT);
-    int load_r1     = !(I->flags & IRCG_R1_IN_RAX);
-
-    switch (I->op) {
-    case IR_NOP:
-    case IR_JMP:
-    case IR_LOOP:
-        return;
-
-    case IR_ALLOCA:
-        irCgAllocTmp(ctx, I->dst, start);
-        return;
-
-    case IR_LOAD:
-        if (spill_dst) irCgAllocTmp(ctx, I->dst, start);
-        irCgAllocTmp(ctx, I->r1, start);            /* address */
-        return;
-
-    case IR_STORE:
-        irCgAllocTmp(ctx, I->dst, start);           /* address */
-        if (load_r1) irCgAllocTmp(ctx, I->r1, start);
-        return;
-
-    case IR_LOAD_DEREF:
-        /* dst = *r1; r1 holds a runtime pointer in its slot. */
-        if (spill_dst) irCgAllocTmp(ctx, I->dst, start);
-        irCgAllocTmp(ctx, I->r1, start);
-        return;
-
-    case IR_STORE_DEREF:
-        /* *dst = r1; dst holds a runtime pointer, r1 the value. */
-        irCgAllocTmp(ctx, I->dst, start);
-        irCgAllocTmp(ctx, I->r1, start);
-        return;
-
-    case IR_LEA:
-        /* dst = &r1's slot. dst gets a slot to hold the pointer value;
-         * r1 is the slot we're taking the address of (already bound
-         * to its loff via param-spill or alloca). */
-        if (spill_dst) irCgAllocTmp(ctx, I->dst, start);
-        return;
-
-    case IR_IADD: case IR_ISUB: case IR_IMUL:
-    case IR_AND:  case IR_OR:   case IR_XOR:
-    case IR_IDIV: case IR_UDIV: case IR_IREM: case IR_UREM:
-    case IR_SHL:  case IR_SHR:  case IR_SAR:
-    case IR_ICMP:
-        if (spill_dst) irCgAllocTmp(ctx, I->dst, start);
-        if (load_r1)   irCgAllocTmp(ctx, I->r1, start);
-        irCgAllocTmp(ctx, I->r2, start);
-        return;
-
-    case IR_FADD: case IR_FSUB: case IR_FMUL: case IR_FDIV:
-    case IR_FCMP:
-        /* Float ops always go through xmm0/xmm1 (no fusion path), so
-         * unconditionally allocate slots for dst and operands. */
-        irCgAllocTmp(ctx, I->dst, start);
-        irCgAllocTmp(ctx, I->r1, start);
-        irCgAllocTmp(ctx, I->r2, start);
-        return;
-
-    case IR_FNEG:
-        irCgAllocTmp(ctx, I->dst, start);
-        irCgAllocTmp(ctx, I->r1, start);
-        return;
-
-    case IR_BR:
-        if (load_r1) irCgAllocTmp(ctx, I->dst, start);
-        return;
-
-    case IR_RET:
-        if (I->dst && load_r1) irCgAllocTmp(ctx, I->dst, start);
-        return;
-
-    case IR_PHI:
-        /* Rax-resident phi: no slot. Slot-resident phi: dst gets a slot,
-         * predecessors store into it, in-block consumers reload from it. */
-        if (!(I->flags & IRCG_PHI_IN_RAX)) {
-            irCgAllocTmp(ctx, I->dst, start);
-        }
-        return;
-
-    case IR_CALL:
-        /* dst (return value) gets a slot like any other def, unless fused.
-         * r1 is the args wrapper (IR_VAL_UNRESOLVED), not a tmp — no slot.
-         * The arg values themselves are loaded directly into argument
-         * registers from their existing slots at emit time.
-         * r2, when set, is the indirect call target — it lives in a slot
-         * just like any other tmp. */
-        if (spill_dst && I->dst && I->dst->type != IR_TYPE_VOID &&
-            I->dst->kind == IR_VAL_TMP) {
-            irCgAllocTmp(ctx, I->dst, start);
-        }
-        irCgAllocTmp(ctx, I->r2, start);
-        return;
-
-    case IR_GEP: {
-        /* Get-element-pointer for stack-allocated structs. dst aliases
-         * `base + offset` — we bind its loff right here so subsequent
-         * loads/stores through it just look up the right %rbp offset. */
-        if (!I->dst || I->dst->kind != IR_VAL_TMP ||
-            !I->r1 || !I->r2 || I->r2->kind != IR_VAL_CONST_INT) {
-            return;
-        }
-        int base_loff = irCgGetLoff(ctx, I->r1);
-        int field_loff = base_loff + (int)I->r2->as._i64;
-        if (!mapHasInt(ctx->id_to_loff, I->dst->as.var.id)) {
-            irCgSetLoff(ctx, I->dst->as.var.id, field_loff);
-        }
-        return;
-    }
-
-    default:
-        /* Be conservative for ops we don't model yet. */
-        irCgAllocTmp(ctx, I->dst, start);
-        irCgAllocTmp(ctx, I->r1, start);
-        irCgAllocTmp(ctx, I->r2, start);
-    }
-}
-
-static void irCgAllocAllTmps(IrCgCtx *ctx, int starting_offset) {
-    listForEach(ctx->func->blocks) {
-        IrBlock *block = (IrBlock *)it->value;
-        listForEach(block->instructions) {
-            IrInstr *instr = (IrInstr *)it->value;
-            irCgAllocOperandsForInstr(ctx, instr, starting_offset);
-        }
-    }
-}
 
 /* Load any IR value into the given 64-bit register (rax or rcx). */
 static void irCgLoadToReg(IrCgCtx *ctx, IrValue *val, const char *reg) {
@@ -307,7 +92,7 @@ static void irCgLoadToReg(IrCgCtx *ctx, IrValue *val, const char *reg) {
     case IR_VAL_LOCAL:
     case IR_VAL_PARAM:
         aoStrCatPrintf(ctx->buf, "movq   %d(%%rbp), %%%s\n\t",
-                       irCgGetLoff(ctx, val), reg);
+                       irCgGetLoff(&ctx->ra, val), reg);
         return;
     default:
         loggerPanic("ir-cg: cannot load value of kind %d\n", val->kind);
@@ -316,7 +101,7 @@ static void irCgLoadToReg(IrCgCtx *ctx, IrValue *val, const char *reg) {
 
 static void irCgStoreReg(IrCgCtx *ctx, IrValue *dst, const char *reg) {
     aoStrCatPrintf(ctx->buf, "movq   %%%s, %d(%%rbp)\n\t",
-                   reg, irCgGetLoff(ctx, dst));
+                   reg, irCgGetLoff(&ctx->ra, dst));
 }
 
 /* Emit `.data <label>: .quad <bits> .text` lazily for a float constant.
@@ -326,7 +111,7 @@ static char *irCgEmitFloatLiteralData(IrCgCtx *ctx, double f) {
     char buf[64];
     snprintf(buf, sizeof(buf), ".LIRF%d_%llu",
              ctx->func_uid,
-             (unsigned long long)ctx->extra_stack);
+             (unsigned long long)ctx->ra.extra_stack);
     /* extra_stack changes per allocation; combined with func_uid that gives
      * us a unique-enough label. The chosen scheme just needs to avoid
      * collisions inside this translation unit. */
@@ -369,7 +154,7 @@ static void irCgLoadToXmm(IrCgCtx *ctx, IrValue *val, const char *xmm) {
     case IR_VAL_LOCAL:
     case IR_VAL_PARAM:
         aoStrCatPrintf(ctx->buf, "movsd  %d(%%rbp), %%%s\n\t",
-                       irCgGetLoff(ctx, val), xmm);
+                       irCgGetLoff(&ctx->ra, val), xmm);
         return;
     default:
         loggerPanic("ir-cg: cannot load float of kind %d\n", val->kind);
@@ -378,7 +163,44 @@ static void irCgLoadToXmm(IrCgCtx *ctx, IrValue *val, const char *xmm) {
 
 static void irCgStoreXmm(IrCgCtx *ctx, IrValue *dst, const char *xmm) {
     aoStrCatPrintf(ctx->buf, "movsd  %%%s, %d(%%rbp)\n\t",
-                   xmm, irCgGetLoff(ctx, dst));
+                   xmm, irCgGetLoff(&ctx->ra, dst));
+}
+
+/* Map a cmp kind to the matching `j[cc]` mnemonic. `is_float` picks
+ * between signed-int (`jl`/`jle`/...) and unsigned/float (`jb`/`jbe`/...)
+ * forms - x86 ucomisd/`cmpq <unsigned>` use the unsigned variants. */
+static const char *irCgJccFor(IrCmpKind cmp, int is_float) {
+    switch (cmp) {
+    case IR_CMP_EQ:  return "je";
+    case IR_CMP_NE:  return "jne";
+    case IR_CMP_LT:  return is_float ? "jb"  : "jl";
+    case IR_CMP_LE:  return is_float ? "jbe" : "jle";
+    case IR_CMP_GT:  return is_float ? "ja"  : "jg";
+    case IR_CMP_GE:  return is_float ? "jae" : "jge";
+    case IR_CMP_ULT: return "jb";
+    case IR_CMP_ULE: return "jbe";
+    case IR_CMP_UGT: return "ja";
+    case IR_CMP_UGE: return "jae";
+    default:         return NULL;
+    }
+}
+
+/* Inverse jcc for "branch if cmp is false". Used when layout has the
+ * true target falling through, so we j[invcc] over to the false arm. */
+static const char *irCgJccInvFor(IrCmpKind cmp, int is_float) {
+    switch (cmp) {
+    case IR_CMP_EQ:  return "jne";
+    case IR_CMP_NE:  return "je";
+    case IR_CMP_LT:  return is_float ? "jae" : "jge";
+    case IR_CMP_LE:  return is_float ? "ja"  : "jg";
+    case IR_CMP_GT:  return is_float ? "jbe" : "jle";
+    case IR_CMP_GE:  return is_float ? "jb"  : "jl";
+    case IR_CMP_ULT: return "jae";
+    case IR_CMP_ULE: return "ja";
+    case IR_CMP_UGT: return "jbe";
+    case IR_CMP_UGE: return "jb";
+    default:         return NULL;
+    }
 }
 
 /* Emit `<setcc> %al; movzbq %al, %rax` for a comparison kind. */
@@ -405,315 +227,13 @@ static void irCgBlockLabel(IrCgCtx *ctx, IrBlock *block, char *out, int n) {
     snprintf(out, n, ".LIRBB%d_%u", ctx->func_uid, block->id);
 }
 
-/* Width in bytes that the codegen should use when reading/writing an
- * IrValue. Tmps carry an explicit `var.size`; other kinds fall back to
- * the IR type tag (pointer/F64/I64 = 8, I32 = 4, I16 = 2, I8/CHAR = 1). */
-static u32 irValueByteSize(IrValue *v) {
-    if (!v) return 8;
-    if (v->kind == IR_VAL_TMP || v->kind == IR_VAL_LOCAL ||
-        v->kind == IR_VAL_PARAM) {
-        if (v->as.var.size > 0) return v->as.var.size;
-    }
-    switch (v->type) {
-    case IR_TYPE_I8:  return 1;
-    case IR_TYPE_I16: return 2;
-    case IR_TYPE_I32: return 4;
-    default:          return 8;
-    }
-}
-
-/* ---- fusion pre-pass --------------------------------------------------- */
-
-/* True if the instruction's natural codegen ends with the result in %rax,
- * i.e. it's a candidate to "leave in rax" rather than spill to a slot.
- * Float-typed defs land in %xmm0, not %rax — they can't ride the fusion
- * path (consumers like IR_RET float / IR_FADD load via movsd from a slot,
- * not rax). Returning 0 keeps the slot-spill in place. */
-static int instrDefsIntoRax(IrInstr *I) {
-    if (I->dst && irIsFloat(I->dst->type)) return 0;
-    if (I->op == IR_PHI && (I->flags & IRCG_PHI_IN_RAX)) return 1;
-    switch (I->op) {
-    case IR_LOAD:
-    case IR_LOAD_DEREF:
-    case IR_LEA:
-    case IR_IADD: case IR_ISUB: case IR_IMUL:
-    case IR_AND:  case IR_OR:   case IR_XOR:
-    case IR_IDIV: case IR_UDIV: case IR_IREM: case IR_UREM:
-    case IR_SHL:  case IR_SHR:  case IR_SAR:
-    case IR_ICMP:
-        return 1;
-    case IR_CALL:
-        /* Non-void return value lives in %rax after the call. */
-        return I->dst && I->dst->type != IR_TYPE_VOID;
-    default:
-        return 0;
-    }
-}
-
-static IrValue *firstRaxSource(IrInstr *I);
-
-/* Classify each IR_PHI: rax-resident if
- *   (a) it's the only phi at the block head,
- *   (b) every predecessor arrives via IR_JMP / IR_LOOP (so %rax survives),
- *   (c) the phi's dst has at most one use across the function, AND
- *   (d) that use is the very next non-NOP instruction in this block, as
- *       its first-rax source — so the value in rax is consumed before any
- *       arithmetic / call clobbers it.
- * All other phis fall back to slot-resident (store on the pred side, load
- * on the use side, just like an alloca). */
-static void irCgClassifyPhis(IrFunction *func) {
-    /* Pre-compute use counts for tmp values across the whole function. We
-     * count r1, r2, dst (for BR/RET — they treat dst as a source), and
-     * phi pair values (consumed at materialisation time). */
-    Map *uses = mapNew(64, &map_uint_to_uint_type);
-#define BUMP_TMP(v) do { \
-    if ((v) && (v)->kind == IR_VAL_TMP) { \
-        u32 _id = (v)->as.var.id; \
-        int _n = mapHasInt(uses, _id) ? (int)(intptr_t)mapGetInt(uses, _id) : 0; \
-        mapAdd(uses, (void *)(u64)_id, (void *)(intptr_t)(_n + 1)); \
-    } \
-} while (0)
-    listForEach(func->blocks) {
-        IrBlock *bb = (IrBlock *)it->value;
-        listForEach(bb->instructions) {
-            IrInstr *I = (IrInstr *)it->value;
-            if (I->op == IR_NOP) continue;
-            if (I->op == IR_CALL && I->r1 && I->r1->as.array.values) {
-                Vec *args = I->r1->as.array.values;
-                for (u64 i = 0; i < args->size; ++i) {
-                    BUMP_TMP((IrValue *)args->entries[i]);
-                }
-            } else {
-                BUMP_TMP(I->r1);
-            }
-            BUMP_TMP(I->r2);
-            if (I->op == IR_BR || I->op == IR_RET ||
-                I->op == IR_STORE_DEREF) {
-                /* IR_STORE_DEREF.dst is the runtime pointer (a real source);
-                 * BR/RET treat dst as the condition / return value. */
-                BUMP_TMP(I->dst);
-            }
-            if (I->op == IR_PHI && I->extra.phi_pairs) {
-                for (u64 i = 0; i < I->extra.phi_pairs->size; ++i) {
-                    IrPair *p = vecGet(IrPair *, I->extra.phi_pairs, i);
-                    BUMP_TMP(p->ir_value);
-                }
-            }
-        }
-    }
-#undef BUMP_TMP
-
-    listForEach(func->blocks) {
-        IrBlock *B = (IrBlock *)it->value;
-        if (listEmpty(B->instructions)) continue;
-
-        int phi_count = 0;
-        IrInstr *the_phi = NULL;
-        listForEach(B->instructions) {
-            IrInstr *I = (IrInstr *)it->value;
-            if (I->op == IR_PHI) { phi_count++; the_phi = I; continue; }
-            if (I->op == IR_NOP) continue;
-            break;
-        }
-        if (phi_count != 1) continue;
-
-        Map *preds = irFunctionGetPredecessors(func, B);
-        if (!preds || preds->size == 0) continue;
-
-        int all_jmp = 1;
-        MapIter *iter = mapIterNew(preds);
-        while (mapIterNext(iter)) {
-            IrBlock *P = (IrBlock *)iter->node->value;
-            IrInstr *term = (IrInstr *)listValue(IrInstr *,
-                                                 listTail(P->instructions));
-            if (!term || (term->op != IR_JMP && term->op != IR_LOOP)) {
-                all_jmp = 0; break;
-            }
-        }
-        mapIterRelease(iter);
-        if (!all_jmp) continue;
-
-        if (!the_phi->dst || the_phi->dst->kind != IR_VAL_TMP) continue;
-        /* Float phis can't ride %rax — the materialisation and consumer go
-         * through xmm0, not rax. Skip. */
-        if (irIsFloat(the_phi->dst->type)) continue;
-        u32 dst_id = the_phi->dst->as.var.id;
-        int dst_uses = mapHasInt(uses, dst_id)
-                       ? (int)(intptr_t)mapGetInt(uses, dst_id) : 0;
-        if (dst_uses > 1) continue;
-
-        /* The single use must be in *this* block as the very next non-NOP
-         * instruction's first-rax source — otherwise rax may be clobbered
-         * before the use. (If the use is in a successor's phi pair, the
-         * predecessor-side materialisation can't read it from rax either.) */
-        IrInstr *next = NULL;
-        int seen_phi = 0;
-        listForEach(B->instructions) {
-            IrInstr *I = (IrInstr *)it->value;
-            if (I == the_phi) { seen_phi = 1; continue; }
-            if (!seen_phi) continue;
-            if (I->op == IR_NOP) continue;
-            if (I->op == IR_PHI) continue;   /* skip sibling phis */
-            next = I;
-            break;
-        }
-        if (!next) continue;
-        IrValue *next_src = firstRaxSource(next);
-        if (!next_src || next_src->kind != IR_VAL_TMP) continue;
-        if (next_src->as.var.id != dst_id) continue;
-
-        the_phi->flags |= IRCG_PHI_IN_RAX;
-    }
-
-    mapRelease(uses);
-}
-
-/* The single source operand the instruction loads into %rax first. NULL
- * when the op doesn't have one (alloca, unconditional jmp, nop, ...). */
-static IrValue *firstRaxSource(IrInstr *I) {
-    switch (I->op) {
-    case IR_STORE:
-    case IR_IADD: case IR_ISUB: case IR_IMUL:
-    case IR_AND:  case IR_OR:   case IR_XOR:
-    case IR_IDIV: case IR_UDIV: case IR_IREM: case IR_UREM:
-    case IR_SHL:  case IR_SHR:  case IR_SAR:
-    case IR_ICMP:
-        return I->r1;
-    case IR_BR:
-    case IR_RET:
-        return I->dst;
-    default:
-        return NULL;
-    }
-}
-
-static void bumpUseIfTmp(Map *uses, IrValue *v) {
-    if (!v || v->kind != IR_VAL_TMP) return;
-    int n = mapHasInt(uses, v->as.var.id)
-            ? (int)(intptr_t)mapGetInt(uses, v->as.var.id) : 0;
-    mapAdd(uses, (void *)(u64)v->as.var.id, (void *)(intptr_t)(n + 1));
-}
-
-/* Walk every instruction once, counting tmp source-uses; then walk again to
- * mark FUSE_TO_NEXT / R1_IN_RAX pairs. Sources for use-counting purposes are:
- * r1, r2, and dst (for IR_BR / IR_RET only — those treat dst as a source). */
-static void irCgAnnotate(IrFunction *func) {
-    Map *uses = mapNew(64, &map_uint_to_uint_type);
-
-    listForEach(func->blocks) {
-        IrBlock *bb = (IrBlock *)it->value;
-        listForEach(bb->instructions) {
-            IrInstr *I = (IrInstr *)it->value;
-            if (I->op == IR_CALL && I->r1 && I->r1->as.array.values) {
-                /* IR_CALL's r1 is the args wrapper — count each arg, not
-                 * the wrapper itself. */
-                Vec *args = I->r1->as.array.values;
-                for (u64 i = 0; i < args->size; ++i) {
-                    bumpUseIfTmp(uses, (IrValue *)args->entries[i]);
-                }
-            } else {
-                bumpUseIfTmp(uses, I->r1);
-            }
-            bumpUseIfTmp(uses, I->r2);
-            if (I->op == IR_BR || I->op == IR_RET ||
-                I->op == IR_STORE_DEREF) {
-                /* IR_STORE_DEREF.dst is the runtime pointer (a real source);
-                 * BR/RET treat dst as the condition / return value. */
-                bumpUseIfTmp(uses, I->dst);
-            }
-            /* Phi pair values are consumed at predecessor-side
-             * materialisation; count them so the zero-use slot-skip
-             * heuristic below doesn't drop slots a phi still reads from. */
-            if (I->op == IR_PHI && I->extra.phi_pairs) {
-                for (u64 i = 0; i < I->extra.phi_pairs->size; ++i) {
-                    IrPair *p = vecGet(IrPair *, I->extra.phi_pairs, i);
-                    bumpUseIfTmp(uses, p->ir_value);
-                }
-            }
-            /* Make sure stale flag bits from a previous compilation can't
-             * leak in if instructions ever get recycled. */
-            I->flags &= ~(u64)(IRCG_FUSE_TO_NEXT | IRCG_R1_IN_RAX);
-        }
-    }
-
-    listForEach(func->blocks) {
-        IrBlock *bb = (IrBlock *)it->value;
-        if (listEmpty(bb->instructions)) continue;
-        for (List *node = bb->instructions->next;
-             node != bb->instructions;
-             node = node->next)
-        {
-            IrInstr *cur = (IrInstr *)node->value;
-            if (cur->op == IR_NOP) continue;
-            if (!instrDefsIntoRax(cur)) continue;
-            if (!cur->dst || cur->dst->kind != IR_VAL_TMP) continue;
-
-            int use_count = mapHasInt(uses, cur->dst->as.var.id)
-                            ? (int)(intptr_t)mapGetInt(uses, cur->dst->as.var.id)
-                            : 0;
-            /* Zero-use def: the result is dead, so we still emit the op
-             * (it may have side effects — IR_CALL — or simply be cheap to
-             * leave) but skip the spill. The slot allocator drops the
-             * unused dst's slot. */
-            if (use_count == 0) {
-                cur->flags |= IRCG_FUSE_TO_NEXT;
-                continue;
-            }
-            if (use_count != 1) continue;
-
-            /* Find the next instruction that actually emits something. Skip
-             * IR_NOPs left by mem2reg / fold. If we hit an IR_JMP whose
-             * target has a single predecessor (this block) we can chase
-             * into the target — %rax survives an unconditional jump. We
-             * stop at IR_BR / IR_LOOP since branch arms may need different
-             * rax setups. */
-            IrInstr *next = NULL;
-            IrBlock *scan_bb = bb;
-            List *scan = node->next;
-            int hops = 0;
-            while (1) {
-                while (scan == scan_bb->instructions) {
-                    /* end of current block — try to chase through a jmp */
-                    IrInstr *term = (IrInstr *)listValue(IrInstr *,
-                                                         listTail(scan_bb->instructions));
-                    if (!term || term->op != IR_JMP) { scan_bb = NULL; break; }
-                    IrBlock *tgt = term->extra.blocks.target_block;
-                    if (!tgt) { scan_bb = NULL; break; }
-                    Map *preds = irFunctionGetPredecessors(func, tgt);
-                    if (!preds || preds->size != 1) { scan_bb = NULL; break; }
-                    scan_bb = tgt;
-                    scan = scan_bb->instructions->next;
-                    if (++hops > 8) { scan_bb = NULL; break; }
-                }
-                if (!scan_bb) break;
-                IrInstr *cand = (IrInstr *)scan->value;
-                if (cand->op != IR_NOP && cand->op != IR_JMP) {
-                    next = cand;
-                    break;
-                }
-                /* Treat IR_JMP as "end of block" so the outer loop chases. */
-                scan = scan->next;
-            }
-            if (!next) continue;
-
-            IrValue *next_src = firstRaxSource(next);
-            if (!next_src || next_src->kind != IR_VAL_TMP) continue;
-            if (next_src->as.var.id != cur->dst->as.var.id) continue;
-
-            cur->flags |= IRCG_FUSE_TO_NEXT;
-            next->flags |= IRCG_R1_IN_RAX;
-        }
-    }
-
-    mapRelease(uses);
-}
 
 /* Emit the body of a single IR instruction. Block boundaries / labels are
  * handled by the caller. */
 /* Helpers wrapping the spill / first-rax-load sites so the bit-flag checks
  * happen in one place. */
 static void irCgLoadFirstSrc(IrCgCtx *ctx, IrInstr *instr, IrValue *src) {
-    if (instr->flags & IRCG_R1_IN_RAX) return;
+    if (instr->flags & IRCG_R1_IN_REG) return;
     irCgLoadToReg(ctx, src, "rax");
 }
 
@@ -732,14 +252,41 @@ static void irCgSpillDst(IrCgCtx *ctx, IrInstr *instr, const char *reg) {
  * slot A is about to read. We collect all phis here, then repeatedly pick
  * one whose dst is not read by any still-pending phi (a leaf in the
  * read-dependency graph). Cycles fall back to scratch via %rcx. */
+/* True when the predecessor block (`from`) ended with a value-producing
+ * instruction whose dst is `v` and which the peephole pass marked
+ * IRCG_FUSE_TO_NEXT (suppressed-spill, "value is in %rax going into
+ * the JMP"). Lets the phi-mat skip the otherwise-redundant slot
+ * reload. */
+static int phiPairValueLiveInRax(IrBlock *from, IrValue *v) {
+    if (!from || !v || v->kind != IR_VAL_TMP) return 0;
+    if (listEmpty(from->instructions)) return 0;
+    for (List *node = from->instructions->prev;
+         node != from->instructions;
+         node = node->prev) {
+        IrInstr *I = (IrInstr *)node->value;
+        if (I->op == IR_NOP) continue;
+        if (I->op == IR_JMP || I->op == IR_BR ||
+            I->op == IR_LOOP || I->op == IR_RET) continue;
+        if (!I->dst || I->dst->kind != IR_VAL_TMP) return 0;
+        if (I->dst->as.var.id != v->as.var.id) return 0;
+        return (I->flags & IRCG_FUSE_TO_NEXT) != 0;
+    }
+    return 0;
+}
+
 static void irCgEmitOnePhi(IrCgCtx *ctx, IrInstr *phi, IrPair *match) {
     IrValue *v = match->ir_value;
+    /* If the predecessor's last def was peephole-tail-fused, %rax (for
+     * int phis) already holds `v` going into this materialise step;
+     * skip both the slot reload and the dangling check below. */
+    int in_rax = !irIsFloat(phi->dst->type)
+                  && phiPairValueLiveInRax(match->ir_block, v);
     /* mem2reg can emit phi pairs whose tmp value flows from a branch
      * that never wrote anything live (typically the first branch of a
      * promoted variable initialised inside the if/else). Such tmps have
      * no slot — treat them as 0/undef. */
-    int v_dangling = v && v->kind == IR_VAL_TMP &&
-                     !mapHasInt(ctx->id_to_loff, v->as.var.id);
+    int v_dangling = !in_rax && v && v->kind == IR_VAL_TMP &&
+                     !mapHasInt(ctx->ra.id_to_loff, v->as.var.id);
     if (irIsFloat(phi->dst->type)) {
         if (v_dangling) {
             aoStrCatPrintf(ctx->buf, "xorpd  %%xmm0, %%xmm0\n\t");
@@ -751,10 +298,10 @@ static void irCgEmitOnePhi(IrCgCtx *ctx, IrInstr *phi, IrPair *match) {
     }
     if (v_dangling) {
         aoStrCatPrintf(ctx->buf, "xorq   %%rax, %%rax\n\t");
-    } else {
+    } else if (!in_rax) {
         irCgLoadToReg(ctx, v, "rax");
     }
-    if (!(phi->flags & IRCG_PHI_IN_RAX)) {
+    if (!(phi->flags & IRCG_PHI_IN_REG)) {
         irCgStoreReg(ctx, phi->dst, "rax");
     }
 }
@@ -828,88 +375,6 @@ static void irCgEmitPhiMaterialize(IrCgCtx *ctx, IrBlock *from, IrBlock *to) {
     }
 }
 
-/* Mirror of the BR/JMP emission's layout decisions: returns the set of
- * block ids that are actually referenced by some asm jump after elision.
- * Blocks not in the set still have their bodies emitted, but their label
- * header is dead code and can be dropped. */
-static Set *irCgComputeReferencedBlocks(IrFunction *func);
-
-/* Does this block start with at least one IR_PHI? */
-static int blockHasPhi(IrBlock *bb) {
-    if (!bb) return 0;
-    listForEach(bb->instructions) {
-        IrInstr *I = (IrInstr *)it->value;
-        if (I->op == IR_PHI) return 1;
-        if (I->op != IR_NOP) return 0;
-    }
-    return 0;
-}
-
-static IrInstr *blockTerminator(IrBlock *bb) {
-    IrInstr *term = NULL;
-    listForEach(bb->instructions) {
-        IrInstr *I = (IrInstr *)it->value;
-        if (I->op == IR_NOP) continue;
-        term = I;
-    }
-    return term;
-}
-
-static Set *irCgComputeReferencedBlocks(IrFunction *func) {
-    Set *referenced = setNew(16, &set_uint_type);
-    for (List *bn = func->blocks->next;
-         bn != func->blocks;
-         bn = bn->next)
-    {
-        IrBlock *block = (IrBlock *)bn->value;
-        IrBlock *next_block = (bn->next != func->blocks)
-                              ? (IrBlock *)bn->next->value
-                              : NULL;
-        IrInstr *term = blockTerminator(block);
-        if (!term) continue;
-
-        switch (term->op) {
-        case IR_JMP:
-        case IR_LOOP: {
-            IrBlock *t = term->extra.blocks.target_block;
-            if (t && t != next_block) {
-                setAdd(referenced, (void *)(u64)t->id);
-            }
-            break;
-        }
-        case IR_BR: {
-            IrBlock *t = term->extra.blocks.target_block;
-            IrBlock *f = term->extra.blocks.fallthrough_block;
-            int t_phi = blockHasPhi(t);
-            int f_phi = blockHasPhi(f);
-            if (!t_phi && !f_phi) {
-                if (next_block == t) {
-                    if (f) setAdd(referenced, (void *)(u64)f->id);
-                } else if (next_block == f) {
-                    if (t) setAdd(referenced, (void *)(u64)t->id);
-                } else {
-                    if (t) setAdd(referenced, (void *)(u64)t->id);
-                    if (f) setAdd(referenced, (void *)(u64)f->id);
-                }
-            } else {
-                /* Phi case always emits `jmp tlbl` after the true-arm
-                 * materialisation; the trailing `jmp flbl` is elidable
-                 * when next == f. */
-                if (t) setAdd(referenced, (void *)(u64)t->id);
-                if (f && next_block != f) {
-                    setAdd(referenced, (void *)(u64)f->id);
-                }
-            }
-            break;
-        }
-        default:
-            /* IR_RET — no jump targets to mark. */
-            break;
-        }
-    }
-    return referenced;
-}
-
 static void irCgEmitInstr(IrCgCtx *ctx, IrInstr *instr) {
     switch (instr->op) {
     case IR_NOP:
@@ -937,7 +402,7 @@ static void irCgEmitInstr(IrCgCtx *ctx, IrInstr *instr) {
          * alloca's 8-byte-aligned slot, but it can also be a GEP-aliased
          * field within a packed class. Pick the load width from the
          * destination type so we don't pull in adjacent-field bytes. */
-        int loff = irCgGetLoff(ctx, instr->r1);
+        int loff = irCgGetLoff(&ctx->ra, instr->r1);
         u32 sz = instr->dst ? instr->dst->as.var.size : 8;
         switch (sz) {
         case 1:
@@ -963,7 +428,7 @@ static void irCgEmitInstr(IrCgCtx *ctx, IrInstr *instr) {
          * the assignment target so this width comes out right for both
          * direct slot stores and IR_GEP-aliased field writes). */
         irCgLoadFirstSrc(ctx, instr, instr->r1);
-        int loff = irCgGetLoff(ctx, instr->dst);
+        int loff = irCgGetLoff(&ctx->ra, instr->dst);
         u32 sz = irValueByteSize(instr->r1);
         switch (sz) {
         case 1:
@@ -1029,7 +494,7 @@ static void irCgEmitInstr(IrCgCtx *ctx, IrInstr *instr) {
             }
             aoStrCatPrintf(ctx->buf, "leaq   %s(%%rip), %%rax\n\t", name);
         } else {
-            int loff = irCgGetLoff(ctx, instr->r1);
+            int loff = irCgGetLoff(&ctx->ra, instr->r1);
             aoStrCatPrintf(ctx->buf, "leaq   %d(%%rbp), %%rax\n\t", loff);
         }
         irCgSpillDst(ctx, instr, "rax");
@@ -1096,6 +561,9 @@ static void irCgEmitInstr(IrCgCtx *ctx, IrInstr *instr) {
         irCgLoadFirstSrc(ctx, instr, instr->r1);
         irCgLoadToReg(ctx, instr->r2, "rcx");
         aoStrCatPrintf(ctx->buf, "cmpq   %%rcx, %%rax\n\t");
+        /* Peephole-fused with the next IR_BR: skip setcc + movzbq +
+         * spill and let the BR turn the flags into a direct j[cc]. */
+        if (instr->flags & IRCG_CMP_FUSED_BR) break;
         irCgEmitSetCC(ctx, instr->extra.cmp_kind);
         irCgSpillDst(ctx, instr, "rax");
         break;
@@ -1207,6 +675,8 @@ static void irCgEmitInstr(IrCgCtx *ctx, IrInstr *instr) {
         irCgLoadToXmm(ctx, instr->r1, "xmm0");
         irCgLoadToXmm(ctx, instr->r2, "xmm1");
         aoStrCatPrintf(ctx->buf, "ucomisd %%xmm1, %%xmm0\n\t");
+        /* Peephole-fused with the next IR_BR. */
+        if (instr->flags & IRCG_CMP_FUSED_BR) break;
         const char *cc = NULL;
         switch (instr->extra.cmp_kind) {
         case IR_CMP_EQ: cc = "sete";  break;
@@ -1231,7 +701,28 @@ static void irCgEmitInstr(IrCgCtx *ctx, IrInstr *instr) {
         irCgBlockLabel(ctx, t, tlbl, sizeof(tlbl));
         irCgBlockLabel(ctx, f, flbl, sizeof(flbl));
 
-        irCgLoadFirstSrc(ctx, instr, instr->dst);
+        /* Peephole-fused with the immediately-prior IR_ICMP / IR_FCMP:
+         * the cmp's flags are still live (it skipped setcc/spill), so
+         * we can branch directly off them instead of the usual
+         * `load-cond-into-rax + testq + je/jne` triple. */
+        int fused_br = (instr->flags & IRCG_BR_USE_PRIOR_CMP) != 0;
+        IrCmpKind fused_kind = IR_CMP_INVALID;
+        int fused_is_float = 0;
+        if (fused_br) {
+            for (List *node = ctx->cur_block->instructions->prev;
+                 node != ctx->cur_block->instructions;
+                 node = node->prev)
+            {
+                IrInstr *prev = (IrInstr *)node->value;
+                if (prev == instr) continue;
+                if (prev->op == IR_NOP) continue;
+                fused_kind = prev->extra.cmp_kind;
+                fused_is_float = (prev->op == IR_FCMP);
+                break;
+            }
+        }
+
+        if (!fused_br) irCgLoadFirstSrc(ctx, instr, instr->dst);
 
         int t_phi = blockHasPhi(t);
         int f_phi = blockHasPhi(f);
@@ -1240,6 +731,21 @@ static void irCgEmitInstr(IrCgCtx *ctx, IrInstr *instr) {
              * fall-through. If the next block is the true target, use
              * `je flbl` and fall through. If it's the false target,
              * invert to `jne tlbl`. Otherwise emit both jumps. */
+            if (fused_br) {
+                const char *jcc_t = irCgJccFor(fused_kind, fused_is_float);
+                const char *jcc_f = irCgJccInvFor(fused_kind, fused_is_float);
+                if (ctx->next_block == t) {
+                    aoStrCatPrintf(ctx->buf, "%s    %s\n\t", jcc_f, flbl);
+                } else if (ctx->next_block == f) {
+                    aoStrCatPrintf(ctx->buf, "%s    %s\n\t", jcc_t, tlbl);
+                } else {
+                    aoStrCatPrintf(ctx->buf,
+                                   "%s    %s\n\t"
+                                   "jmp    %s\n\t",
+                                   jcc_t, tlbl, flbl);
+                }
+                break;
+            }
             aoStrCatPrintf(ctx->buf, "testq  %%rax, %%rax\n\t");
             if (ctx->next_block == t) {
                 aoStrCatPrintf(ctx->buf, "je     %s\n\t", flbl);
@@ -1264,10 +770,15 @@ static void irCgEmitInstr(IrCgCtx *ctx, IrInstr *instr) {
             int br_id = br_seq++;
             snprintf(else_lbl, sizeof(else_lbl), ".LIRBR%d_E%d",
                      ctx->func_uid, br_id);
-            aoStrCatPrintf(ctx->buf,
-                           "testq  %%rax, %%rax\n\t"
-                           "je     %s\n\t",
-                           else_lbl);
+            if (fused_br) {
+                const char *jcc_f = irCgJccInvFor(fused_kind, fused_is_float);
+                aoStrCatPrintf(ctx->buf, "%s    %s\n\t", jcc_f, else_lbl);
+            } else {
+                aoStrCatPrintf(ctx->buf,
+                               "testq  %%rax, %%rax\n\t"
+                               "je     %s\n\t",
+                               else_lbl);
+            }
             irCgEmitPhiMaterialize(ctx, ctx->cur_block, t);
             aoStrCatPrintf(ctx->buf, "jmp    %s\n", tlbl);
             asmRemovePreviousTab(ctx->buf);
@@ -1506,119 +1017,6 @@ static void irCgEmitInstr(IrCgCtx *ctx, IrInstr *instr) {
  * left and don't deserve a slot. Slice constraints (no varargs, int-only
  * locals/params, AST_LVAR params) keep this short. Returns the locals
  * + params reservation, NOT 16-byte-aligned. */
-static int irCgComputeAstLayout(Ast *ast_func, IrFunction *ir_func) {
-    Set *surviving = setNew(8, &set_uint_type);
-    listForEach(ir_func->blocks) {
-        IrBlock *bb = (IrBlock *)it->value;
-        listForEach(bb->instructions) {
-            IrInstr *I = (IrInstr *)it->value;
-            if (I->op == IR_ALLOCA && I->dst &&
-                I->dst->kind == IR_VAL_TMP) {
-                setAdd(surviving, (void *)(u64)I->dst->as.var.id);
-            }
-        }
-    }
-
-    int total = 0;
-    int new_offset = 0;
-
-    listForEach(ast_func->locals) {
-        Ast *l = (Ast *)it->value;
-        u32 lid;
-        int sz;
-        if (l->kind == AST_DEFAULT_PARAM) {
-            lid = l->declvar->lvar_id;
-            sz = l->declvar->type->size;
-        } else if (l->kind == AST_FUNPTR) {
-            lid = l->fn_ptr_id;
-            sz = 8; /* function pointer is always 8 bytes */
-        } else {
-            lid = l->lvar_id;
-            sz = l->type->size;
-        }
-        IrValue *iv = irFnGetVar(ir_func, lid);
-        int promoted = iv && iv->kind == IR_VAL_TMP &&
-                       !setHas(surviving, (void *)(u64)iv->as.var.id);
-        if (promoted) {
-            /* mem2reg killed the alloca — skip the slot entirely. The AST
-             * loff is no longer reachable (loads/stores via lvar_id were
-             * NOP'd) so leaving it 0 is safe. */
-            continue;
-        }
-        total += align(sz, 8);
-        new_offset -= sz;
-        l->loff = new_offset;
-    }
-
-    int param_total = 0;
-    AstType *rt = ast_func->type ? ast_func->type->rettype : NULL;
-    int has_hidden_out_ptr = rt &&
-        (rt->kind == AST_TYPE_CLASS || rt->kind == AST_TYPE_UNION) &&
-        !rt->is_intrinsic && rt->size > 0;
-    if (has_hidden_out_ptr) {
-        /* Reserve 8 bytes for the hidden out-pointer (rdi spill). */
-        param_total += 8;
-    }
-    if (ast_func->params) {
-        for (u64 i = 0; i < ast_func->params->size; ++i) {
-            Ast *p = vecGet(Ast *, ast_func->params, i);
-            int sz;
-            if (p->kind == AST_FUNPTR) {
-                sz = 8;
-            } else if (p->kind == AST_VAR_ARGS) {
-                /* Only argc needs a slot. argv directly aliases the
-                 * caller's variadic area at +16(%rbp) — no spill. */
-                sz = 8;
-            } else {
-                sz = p->type->size;
-            }
-            param_total += align(sz, 8);
-        }
-    }
-    total += param_total;
-    int locals_aligned = total ? align(total, 16) : 0;
-
-    /* Params live at the bottom of the locals+params region: first param
-     * at the most-negative offset, working upward toward locals. Matches
-     * asmFunctionInit's layout. The hidden out-pointer (struct return)
-     * is the very first param - takes the deepest slot, and bind its
-     * loff to the IR-side IrValue (which lives on `ir_func->return_value`
-     * for class-returning functions). */
-    int offset = locals_aligned;
-    if (has_hidden_out_ptr) {
-        /* Stash the loff inside the IR_VAL_PARAM's `var.size` upper
-         * bits? No - simpler: store it temporarily in an out-of-band
-         * field. But IrValue has none free. Use the AST function's
-         * `loff` field as a side channel - it's otherwise unused for
-         * AST_FUNC nodes. The codegen prologue + irCgBindAstLoffs
-         * read it back. */
-        ast_func->loff = -offset;
-        offset -= 8;
-    }
-    if (ast_func->params) {
-        for (u64 i = 0; i < ast_func->params->size; ++i) {
-            Ast *p = vecGet(Ast *, ast_func->params, i);
-            if (p->kind == AST_VAR_ARGS) {
-                /* argc gets a real slot (carries the count register).
-                 * argv has type AST_TYPE_ARRAY in the AST, so the body
-                 * lowers `argv[i]` as `*(&argv + i*sizeof(elem))` —
-                 * meaning `&argv` already produces an absolute address.
-                 * Point its loff at +16(%rbp), which is exactly where
-                 * the caller pushed the variadic values. No bytes
-                 * reserved for it inside this frame. */
-                p->argc->loff = -offset;
-                offset -= 8;
-                p->argv->loff = 16;
-                continue;
-            }
-            int sz = (p->kind == AST_FUNPTR) ? 8 : p->type->size;
-            p->loff = -offset;
-            offset -= align(sz, 8);
-        }
-    }
-    setRelease(surviving);
-    return locals_aligned;
-}
 
 static void irCgEmitFunctionPrologue(AoStr *buf, Ast *func, int total_stack) {
     char *fname = asmNormaliseFunctionName(func->fname->data);
@@ -1693,6 +1091,10 @@ void asmFunctionFromIr(Cctrl *cc, AoStr *buf, Ast *ast_func) {
     irFoldFunction(func);
     irCgClassifyPhis(func);
     irCgAnnotate(func);
+    /* Peephole runs *after* annotation: the annotation pass clears
+     * IRCG_FUSE_TO_NEXT / IRCG_R1_IN_REG before its own re-marking,
+     * so any flag the peephole sets earlier would be wiped. */
+    irPeephole(func);
 
     /* 3. Compute AST-side layout *after* mem2reg so promoted locals
      *    don't reserve dead slots. */
@@ -1702,19 +1104,19 @@ void asmFunctionFromIr(Cctrl *cc, AoStr *buf, Ast *ast_func) {
     IrCgCtx ctx;
     ctx.cc = cc;
     ctx.buf = buf;
-    ctx.func = func;
-    ctx.id_to_loff = mapNew(32, &map_uint_to_uint_type);
-    ctx.extra_stack = 0;
+    ctx.ra.func = func;
+    ctx.ra.id_to_loff = mapNew(32, &map_uint_to_uint_type);
+    ctx.ra.extra_stack = 0;
     ctx.func_uid = ircg_func_seq++;
     ctx.cur_block = NULL;
     ctx.next_block = NULL;
-    irCgBindAstLoffs(&ctx, ast_func);
+    irCgBindAstLoffs(&ctx.ra, ast_func);
 
     /* 5. Allocate slots for IR tmps starting just below the params area. */
-    irCgAllocAllTmps(&ctx, locals_params_space);
+    irCgAllocAllTmps(&ctx.ra, locals_params_space);
 
     /* 6. Emit prologue with one subq covering params + locals + tmps. */
-    int total_stack = locals_params_space + ctx.extra_stack;
+    int total_stack = locals_params_space + ctx.ra.extra_stack;
     if (total_stack > 0) total_stack = align(total_stack, 16);
     irCgEmitFunctionPrologue(buf, ast_func, total_stack);
     irCgEmitParamSpills(buf, ast_func);
@@ -1752,7 +1154,7 @@ void asmFunctionFromIr(Cctrl *cc, AoStr *buf, Ast *ast_func) {
         asmFunctionLeave(buf);
     }
 
-    mapRelease(ctx.id_to_loff);
+    mapRelease(ctx.ra.id_to_loff);
     free(ir_ctx->prog);
     free(ir_ctx);
 }
